+++ CHANGES.md
@@ -1,7 +1,7 @@
 <!-- markdownlint-configure-file { "no-duplicate-heading": { "siblings_only": true } } -->
 
 <!-- markdownlint-disable-next-line first-line-h1 -->
-## 0.4.3 - Unreleased
+## 0.4.3 - 2024-02-28
 
 ### Fixed